Changed: Adapt to the type change of attribute amount from float to double.

+++ app/src/main/java/com/example/wellfed/shoppingcart/ShoppingCartIngredientEditActivity.java
@@ -118,7 +118,7 @@ public class ShoppingCartIngredientEditActivity extends ActivityBase
         }
         shoppingCartIngredient.setDescription(description.getText().toString());
         shoppingCartIngredient.setCategory(category.getText().toString());
-        shoppingCartIngredient.setAmount(Float.parseFloat(amount.getText().toString()));
+        shoppingCartIngredient.setAmount(Double.parseDouble(amount.getText().toString()));
         shoppingCartIngredient.setUnit(unit.getText().toString());
 
         Intent intent = new Intent();
